Make references to atom explicit, not referred
After extensive discussion in https://github.com/reagent-project/reagent/issues/161 it appears a number of new people get confused by the use of the referred reagent atom. Changing this page (and other parts of the documentation) to use an aliased namespace directly without using a referred version of atom may decrease the number of confused people. This is largely an experiment to see if it helps. (It can't hurt!) :)