Proof-of-concept of synthetic input support.
Synthetic inputs are React components that handle input,
but are not literally <input> nodes at their root level.
A synthetic component may be rendered, for example, as a
<div> containing an <input> in some arbitrary sub-location.

Adds support for a :synthetic-input keyword parameter to
adapt-react-class, which tags the component as a recognized
input component. The options map configures hooks for :on-update
and :on-change, to allow overriding default behaviour with
code that is specific to the nuances of the custom component's
DOM. Also adds :on-write keyword parameter to input-node-set-value
to override the value-setting behaviour.