Update ManagingState.md
I actually wanted to log state changes and I was happy to find this example but I think `run!` is a better fit for this purpose. I think it's worth mentioning. IIUC there's nothing to dispose in the `run!` example and no way to stop it. Maybe that's a reason to not bring it up or just something that should get added. Idk.
Thanks for the docs!